\name{biogrid}
\alias{biogrid}
\docType{data}
\title{
  BIOGRID database
}
\description{
BioGrid: Biological General Repository for Interaction Datasets. 515421 unique links.
  (Downloaded from http://thebiogrid.org/ by January 2015
}
\usage{data("biogrid")}

\details{
}
\source{
}
\references{ 
1. Stark C, Breitkreutz BJ, Reguly T, Boucher L, Breitkreutz A, Tyers M: BioGRID: a general repository for interaction datasets. Nucleic Acids Research 2006, 34:D535-D539.
}
\examples{
  
 # see ganet.combine the example in function
  
}
\keyword{databases}
