\name{corum}
\alias{corum}
\docType{data}
\title{
  CORUM database
}
\description{
CORUM: the Comprehensive Resource of Mammalian protein complexes. 44666 unique links.
(Downloaded from http://mips.helmholtz-muenchen.de/genre/proj/corum by January 2015
}
\usage{data("corum")}

\details{
}
\source{
}
\references{ 
1. Ruepp A, Waegele B, Lechner M, Brauner B, Dunger-Kaltenbach I, Fobo G, Frishman G, Montrone C, Mewes HW: CORUM: the comprehensive resource of mammalian protein complexes-2009. Nucleic Acids Research 2010, 38:D497-D501.

}
\examples{
  
 # see ganet.combine the example in function
  
}
\keyword{databases}
