\name{dip}
\alias{dip}
\docType{data}
\title{
  DIP database
}
\description{
DIP: Database of Interacting Proteins. 37945 unique links.
(Downloaded from http://dip.doe-mbi.ucla.edu/dip/Main.cgi by January 2015)
}
\usage{data("dip")}

\details{
}
\source{
}
\references{ 
1.Salwinski L, Miller CS, Smith AJ, Pettit FK, Bowie JU, Eisenberg D: The Database of Interacting Proteins: 2004 update. Nucleic Acids Research 2004, 32:D449-D451.
}
\examples{
  
 # see ganet.combine the example in function
  
}
\keyword{databases}
