\name{hprd}
\alias{hprd}
\docType{data}
\title{
  HPRD database
}
\description{
Human Protein Reference Database (HPRD).
44376 unique links. Downloaded from http://www.hprd.org/download  by January 2015
}
\usage{data("hprd")}

\details{
}
\source{
}
\references{
1. Prasad TSK, Goel R, Kandasamy K, Keerthikumar S, Kumar S, Mathivanan S, Telikicherla D, Raju R, Shafreen B, Venugopal A, et al: Human Protein Reference Database-2009 update. Nucleic Acids Research 2009, 37:D767-D772.
}
\examples{
  
 # see ganet.combine the example in function
  
}
\keyword{databases}
