\name{innatedb}
\alias{innatedb}
\docType{data}
\title{
  innatedb database
}
\description{
A knowledge resource for innate immunity interactions & pathways. 16665 unique links. 
Downloaded from http://www.innatedb.com/redirect.do?go=downloadImported  by January 2015
}
\usage{data("innatedb")}

\details{
}
\source{
}
\references{
1. Breuer K, Foroushani AK, Laird MR, Chen C, Sribnaia A, Lo R, Winsor GL, Hancock REW, Brinkman FSL, Lynn DJ: InnateDB: systems biology of innate immunity and beyond-recent updates and continuing curation. Nucleic Acids Research 2013, 41:D1228-D1233.
}
\examples{
  
 # see ganet.combine the example in function
  
}
\keyword{databases}
