\name{intact}
\alias{intact}
\docType{data}
\title{
  intact database
}
\description{
IntAct provides protein interaction data. 136605 unique links. 
Downloaded from https://www.ebi.ac.uk/intact/ by January 2015
}
\usage{data("intact")}

\details{
}
\source{
}
\references{ 
1.Kerrien S, Aranda B, Breuza L, Bridge A, Broackes-Carter F, Chen C, Duesbury M, Dumousseau M, Feuermann M, Hinz U, et al: The IntAct molecular interaction database in 2012. Nucleic Acids Research 2012, 40:D841-D846.
}
\examples{
  
 # see ganet.combine the example in function
  
}
\keyword{databases}
