\name{mint}
\alias{mint}
\docType{data}
\title{
  mint database
}
\description{
Molecular INTeraction database (MINT). 296806  links. 
Downloaded from http://www.pathwaycommons.org by Nov 2011
}
\usage{data("mint")}

\details{
}
\source{
}
\references{ 
Licata L, Briganti L, Peluso D, Perfetto L, Iannuccelli M, Galeota E, Sacco F, Palma A, Nardozza AP, Santonico E, et al: MINT, the molecular interaction database: 2012 update. Nucleic Acids Research 2012, 40:D857-D861.
}
\examples{
  
 # see ganet.combine the example in function
  
}
\keyword{databases}
