\name{mpi}
\alias{mpi}
\docType{data}
\title{
  MIPS database
}
\description{
MIPS Mammalian Protein-Protein Interaction Database. 807 unique links.
Downloaded from http://mips.helmholtz-muenchen.de/proj/ppi/ by  January 2015
}
\usage{data("mpi")}

\details{
}
\source{
}
\references{ 
1. Pagel P, Kovac S, Oesterheld M, Brauner B, Dunger-Kaltenbach I, Frishman G, Montrone C, Mark P, Stumpflen V, Mewes HW, et al: The MIPS mammalian protein-protein interaction database. Bioinformatics 2005, 21:832-834.
}
\examples{
  
 # see ganet.combine the example in function
  
}
\keyword{databases}
