\name{nat}
\alias{nat}
\docType{data}
\title{
  nat database
}
\description{
Nature Pathway Interaction Database. 12944 links. 
Downloaded from http://www.pathwaycommons.org by Nov 2011,
or the main source is: The Pathway Interaction Database (PID, http://pid.nci.nih.gov)
}
\usage{data("nat")}

\details{
}
\source{
}
\references{ 
1.Schaefer CF, Anthony K, Krupa S, Buchoff J, Day M, Hannay T, Buetow KH. PID: the Pathway Interaction Database. Nucleic Acids Res. 2009;37(database issue):d674-9 
}
\examples{
  
 # see ganet.combine the example in function
  
}
\keyword{databases}
