\name{reactome}
\alias{reactome}
\docType{data}
\title{
  reactome database
}
\description{
Reactome: a knowledgebase of biological processes. 
642272 links. Downloaded from http://www.reactome.org/pages/download-data/ by January  2015
}
\usage{data("reactome")}

\details{
}
\source{
}
\references{
1. Croft D, Mundo AF, Haw R, Milacic M, Weiser J, Wu G, Caudy M, Garapati P, Gillespie M, Kamdar MR, Jassal B, Jupe S, Matthews L, May B, Palatnik S, Rothfels K, Shamovsky V, Song H, Williams M, Birney E, Hermjakob H, Stein L, D'Eustachio P. The Reactome pathway knowledgebase. Nucleic Acids Res. 2014;42(database issue):d472-7
2.Cerami EG, Gross BE, Demir E, Rodchenkov I, Babur O, Anwar N, Schultz N, Bader GD, Sander C: Pathway Commons, a web resource for biological pathway data. Nucleic Acids Research 2011, 39:D685-D690.
}
\examples{
  
 # see ganet.combine the example in function
  
}
\keyword{databases}
